+++ src/protocols/LMS64CProtocol.h
@@ -175,8 +175,8 @@ public:
 
     virtual int ProgramWrite(const char *buffer, const size_t length, const int programmingMode, const int device, ProgrammingCallback callback = 0);
 
-    virtual int CustomParameterRead(const uint8_t *ids, double *values, const int count, std::string* units);
-    virtual int CustomParameterWrite(const uint8_t *ids, const double *values, const int count, const std::string* units);
+    virtual int CustomParameterRead(const uint8_t *ids, double *values, const size_t count, std::string* units);
+    virtual int CustomParameterWrite(const uint8_t *ids, const double *values, const size_t count, const std::string* units);
 
 private:
